requirements.txt
Django==6.0.1
djangorestframework==3.16.1
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
celery==5.6.2
redis==7.1.0
drf-spectacular==0.29.0
django-redis==6.0.0